requirements.txt
Django==4.1.5
boto3==1.26.47
django_storages==1.13.2
django-summernote==0.8.20.0
django-recaptcha==3.0.0
Pillow==9.4.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-cleanup==6.0.0